requirements.txt
discord.py
rich
motor
aiohttp
python-dotenv
pyyaml